+++ Sources/VComponents/Components/Buttons/Loading Stretched Button/VLoadingStretchedButtonUIModel.swift
@@ -147,6 +147,12 @@ public struct VLoadingStretchedButtonUIModel {
         /// Indicates if button animates state change. Defaults to `true`.
         public var animatesStateChange: Bool = true
         
+        /// Ratio to which label scales down on press. Set to `1`.
+        public var backgroundPressedScale: CGFloat = GlobalUIModel.Buttons.pressedScale
+        
+        /// Ratio to which label scales down on press. Set to `1`.
+        public var labelPressedScale: CGFloat = GlobalUIModel.Buttons.pressedScale
+        
         /// Model for customizing spinner animations.
         public var spinnerSubUIModel: VContinuousSpinnerUIModel.Animations = .init()