+++ Sources/VComponents/Components/Buttons/Rounded Labeled Button/VRoundedCaptionedButtonUIModel.swift
@@ -174,6 +174,15 @@ public struct VRoundedCaptionedButtonUIModel {
         /// Indicates if button animates state change. Defaults to `true`.
         public var animatesStateChange: Bool = true
         
+        /// Ratio to which label scales down on press. Set to `1`.
+        public var backgroundPressedScale: CGFloat = GlobalUIModel.Buttons.pressedScale
+        
+        /// Ratio to which label scales down on press. Set to `1`.
+        public var labelPressedScale: CGFloat = GlobalUIModel.Buttons.pressedScale
+        
+        /// Ratio to which caption scales down on press. Set to `1`.
+        public var captionPressedScale: CGFloat = GlobalUIModel.Buttons.pressedScale
+        
         // MARK: Initializers
         /// Initializes UI model with default values.
         public init() {}